feat: move oob nedpoints to own endpoints
